+++ swak/cloud/aws/__init__.py
@@ -9,15 +9,13 @@ from .bucket import S3Bucket
 from .df2s3 import DataFrame2S3Parquet
 from .s32df import S3Parquet2DataFrame
 from .s32local import S3File2LocalFile
+from .discovery import S3ObjectDiscovery
 
 __all__ = [
     'S3',
     'S3Bucket',
     'DataFrame2S3Parquet',
     'S3Parquet2DataFrame',
-    'S3File2LocalFile'
+    'S3File2LocalFile',
+    'S3ObjectDiscovery'
 ]
-
-# ToDo: Add discover files!
-# ToDo: Add download directory to local?
-# ToDo: Add download directory to dataframe?