+++ tests/cloud/aws/test_discovery.py
@@ -1,0 +1,12 @@
+import unittest
+import pickle
+from unittest.mock import patch, Mock
+from pathlib import Path
+from tempfile import TemporaryDirectory
+from swak.cloud.aws.exceptions import S3Error
+from swak.cloud.aws import S3ObjectDiscovery, S3
+
+# ToDO: Continue here!
+
+if __name__ == '__main__':
+    unittest.main()