+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added S3Bucket creation
 - Added overwrite and skip to DataFrame2S3Bucket
 - Added download S3 file to local
+- List (and filter) S3 object keys
 
 ## [0.5.3] - 2025-06-22
 - Added Fallback to funcflow